// frontend/src/app/models/appearance-request.ts
export class AppearanceRequest {
    firstName: string = '';
    lastName: string = '';
    email: string = '';
    phoneNumber: string = '';
    date: string = '';
    startTime: string = '';
    endTime: string = '';
    eventTitle: string = '';
    organization: string = '';
    location: string = '';
    members: string = '';
    specialInstruction: string = '';
    expenses: string = '';
    outsideOrg: string = '';
    description: string = '';
    requiresPerformance: string = '';
    onCampus: boolean;
}